Work Session 4
My main accomplishments this work session was completing project Euler problem's 6 and 8. Personally, the most confusing aspect of number 6 was the fact that I had to code the sum of the squares and the squared sum of numbers 1 to 100, which seems like a very high amount. However, After 5 or 10 minutes of thinking and brainstorming, I realized that putting a for loop inside of my definition of the variables s and s2 could solve the problem. Tt allowed me to process each number, from 1 to 100, while writing an extremely simple code. For problem 8, I initially had the idea of making the 1000-digit number a number value in python. However, I hit multiple roadblocks and did not how to code the problem when the 1000-digit number was classified as a number. Instead, I made the number a string, which allowed me to use the ':' operation tool to differentiate and constantly go through different number combinations in the string.