Work Session 5 Journal
Because I only worked for around 45 minutes, my main accomplishment this work session was completing problem 9. Because the visual studio code terminal can brute force a problem unlike a human, I took advantage of that through for loops. I initially created a for loop that ranged from 1 - 1001 to test the what the variable a could be. To make them correlated, I added another for loop within the previous one that focused on variable b, also ranging from 1 -1001. The for loop allowed me to check every number from 1 to 1000 between variables a and b. After they were decided, I used the two values to find c and tested that they fit into the Pythagorus theorem.