Work Session 6
Although it is not a lot of code, problem 4 took me a long time to figure out. My main issue was coding my if statement to test if the multiplied three digit number was in fact the largest palindromic. At first, I tried looking up a lot of different python tools and functions that would help me with writing a code that tests if the first part of the number sequence is equal to the last part (reversed). After thinking and researching, I finally realized that I could use and index function and an if statement. In order to do this, I changed the multiplied number into a string to separate it and compare each side. After that, I was able to run the code and finish the problem!